Create README.md
About the project

Background: Several laboratory test methods for measuring the chloride diffusion coefficient of concrete have been proposed. The Nordic standard NT Build 492 is one of the accelerated test procedures for the chloride diffusion coefficient, in which chlorides penetrate the concrete at high speed owing to the used electric field. Although the NT Build 492 test technique provides quick results, the test is typically conducted 28 days after the first test concrete mix is made, as this requires curing. If the first attempt is unsuccessful, the whole process will be repeated and this process will consume more time and resources. The test also requires the help of an experienced operator. All these problems complicate the experimental estimation of the diffusion coefficient of concrete for each project. In another perspective, the conventional simplified models cause the material engineers to overdesign the chloride resistance properties of concrete as a precautionary measure, fearing that they will not provide the desired properties. Therefore, the development of accurate models that determine the chloride diffusion coefficients of concrete, taking into account all influencing parameters, is of crucial importance.

Aim: The aim of the project is to develop a reliable, universal, and robust model to predict the chloride migration coefficient, which will eventually replace the cumbersome, time-consuming, and resource-intensive laboratory test.

Objectives: The objectives are twofold: i) developing a machine learning based model to predict the chloride migration coefficient, which takes all influencing parameters into account with a state-of-the-art algorithm and ii) investigating the significance of mix components, fresh and hardened concrete tests to predict chloride migration coefficients.

Data: A comprehensive database of 843 experiments investigating the non-steady-state migration coefficients (𝐷𝑛𝑠𝑠𝑚)
of different concrete types is created by collecting from: i) research projects, and ii) internationally published journal articles. The database stored information about the concrete mix, including its fresh and hardened properties.